+++ graphein/ml/datasets/foldcomp_dataset.py
@@ -11,7 +11,16 @@ import os
 import random
 import shutil
 from pathlib import Path
-from typing import Any, Callable, Dict, Iterable, List, Optional, Union
+from typing import (
+    Any,
+    Callable,
+    Dict,
+    Iterable,
+    List,
+    Literal,
+    Optional,
+    Union,
+)
 
 import numpy as np
 import torch
@@ -58,11 +67,13 @@ except ImportError:
     log.warning(message)
 
 FOLDCOMP_DATABASE_TYPES: List[str] = [
-    "afdb_swissprot_v4",
-    "afdb_uniprot_v4",
-    "afdb_rep_dark_v4",
-    "highquality_clust30",
-    "afdb_rep_v4",
+    "afdb_swissprot_v4",  # AlphaFoldDB Swiss-Prot
+    "afdb_uniprot_v4",  # AlphaFoldDB Uniprot
+    "afdb_rep_v4",  # AlphaFoldDB Cluster Representatives
+    "afdb_rep_dark_v4",  # AlphaFoldDB Cluster Representatives (Dark Clusters)
+    "esmatlas",  # ESMAtlas full (v0 + v2023_02)
+    "esmatlas_v2023_02",  # ESMAtlas v2023_02
+    "highquality_clust30",  # ESMAtlas high-quality
 ]
 """
 Currently supported FoldComp databases. See:
@@ -129,7 +140,15 @@ class FoldCompDataset(Dataset):
     def __init__(
         self,
         root: str,
-        database: str,
+        database: Literal[
+            "afdb_swissprot_v4",
+            "afdb_uniprot_v4",
+            "afdb_rep_v4",
+            "afdb_rep_dark_v4",
+            "esmatlas",
+            "esmatlas_v2023_02",
+            "highquality_clust30",
+        ],
         ids: Optional[List[str]] = None,
         exclude_ids: Optional[List[str]] = None,
         fraction: float = 1.0,
@@ -142,7 +161,7 @@ class FoldCompDataset(Dataset):
         :type root: str
         :param database: Name of the database. See:
             :const:`FOLDCOMP_DATABASE_TYPES`.
-        :type database: str
+        :type database: Literal
         :param ids: List of protein IDs to include in the dataset. If ``None``,
             all proteins are included. Default is ``None``.
         :type ids: Optional[List[str]]
@@ -204,7 +223,9 @@ class FoldCompDataset(Dataset):
     def download(self):
         """Downloads foldcomp database if not already downloaded."""
 
-        if not all(os.path.exists(self.root / f) for f in self._database_files):
+        if not all(
+            os.path.exists(self.root / f) for f in self._database_files
+        ):
             log.info(f"Downloading FoldComp dataset {self.database}...")
             curr_dir = os.getcwd()
             os.chdir(self.root)
@@ -215,10 +236,6 @@ class FoldCompDataset(Dataset):
                 asyncio.run(_)
             os.chdir(curr_dir)
             log.info("Download complete.")
-            # log.info("Moving files to raw directory...")
-
-            # for f in self._database_files:
-            #    shutil.move(f, self.root)
         else:
             log.info(f"FoldComp database already downloaded: {self.root}.")
 
@@ -249,7 +266,9 @@ class FoldCompDataset(Dataset):
             ]
         # Sub sample
         log.info(f"Sampling fraction: {self.fraction}...")
-        accessions = random.sample(accessions, int(len(accessions) * self.fraction))
+        accessions = random.sample(
+            accessions, int(len(accessions) * self.fraction)
+        )
         self.ids = accessions
         log.info("Creating index...")
         indices = dict(enumerate(accessions))
@@ -382,7 +401,9 @@ class FoldCompLightningDataModule(L.LightningDataModule):
         self.val_split = val_split
         self.test_split = test_split
         self.transform = (
-            self._compose_transforms(transform) if transform is not None else None
+            self._compose_transforms(transform)
+            if transform is not None
+            else None
         )
 
         if (
@@ -405,9 +426,12 @@ class FoldCompLightningDataModule(L.LightningDataModule):
             return T.Compose(transforms)
 
     def setup(self, stage: Optional[str] = None):
-        self.train_dataset()
-        self.val_dataset()
-        self.test_dataset()
+        if stage == "fit" or stage is None:
+            self.train_dataset()
+        elif stage == "validate":
+            self.val_dataset()
+        elif stage == "test":
+            self.test_dataset()
 
     def _get_indices(self):
         """Loads the whole database to extract the indices."""
@@ -421,7 +445,9 @@ class FoldCompLightningDataModule(L.LightningDataModule):
         self.ids = ds.ids
         ds.db.close()
 
-    def _split_data(self, train_split: float, val_split: float, test_split: float):
+    def _split_data(
+        self, train_split: float, val_split: float, test_split: float
+    ):
         """Split the database into non-overlapping train, validation and test"""
         if not hasattr(self, "ids"):
             self._get_indices()